+++ reconcile/openshift_groups.py
@@ -28,7 +28,6 @@ from reconcile.typed_queries.clusters import get_clusters
 from reconcile.utils import (
     expiration,
     gql,
-    metrics,
 )
 from reconcile.utils.defer import defer
 from reconcile.utils.oc_map import (
@@ -36,7 +35,6 @@ from reconcile.utils.oc_map import (
     OCMap,
     init_oc_map_from_clusters,
 )
-from reconcile.utils.openshift_resource import OpenshiftResourceInventoryGauge
 from reconcile.utils.secret_reader import create_secret_reader
 from reconcile.utils.sharding import is_in_shard
 
@@ -267,24 +265,6 @@ def act(diff: Mapping[str, Optional[str]], oc_map: ClusterMap) -> None:
         raise Exception("invalid action: {}".format(action))
 
 
-def get_state_count_combinations(state: Iterable[Mapping[str, str]]) -> Counter[str]:
-    return Counter(s["cluster"] for s in state)
-
-
-def publish_metrics(state: Iterable[Mapping[str, str]]) -> None:
-    for cluster, count in get_state_count_combinations(state).items():
-        metrics.set_gauge(
-            OpenshiftResourceInventoryGauge(
-                integration=QONTRACT_INTEGRATION,
-                cluster=cluster,
-                namespace="cluster",
-                kind="Group",
-                state="desired",
-            ),
-            count,
-        )
-
-
 @defer
 def run(
     dry_run: bool,
@@ -312,7 +292,9 @@ def run(
         if not (s["cluster"] in ocm_clusters and s["group"] == "dedicated-admins")
     ]
 
-    publish_metrics(groups_list)
+    ob.publish_cluster_desired_metrics_from_state(
+        groups_list, QONTRACT_INTEGRATION, "Group"
+    )
     diffs = calculate_diff(current_state, desired_state)
     validate_diffs(diffs)
     diffs.sort(key=sort_diffs)
